% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.gen.R
\name{grid.gen}
\alias{grid.gen}
\title{grid.gen: generate grid for ross-validation function}
\usage{
grid.gen(
  y,
  X,
  study = NA,
  grid,
  commonSupp = FALSE,
  lambda_1 = TRUE,
  lambda_2 = FALSE,
  lambda_z = TRUE,
  multiTask = TRUE
)
}
\arguments{
\item{y}{A numeric vector}

\item{X}{A matrix}

\item{study}{An integer vector}

\item{grid}{A dataframe}

\item{commonSupp}{A boolean}

\item{multiTask}{A boolean}

\item{nfold}{An integer}

\item{maxIter}{An integer}

\item{LocSrch_skip}{An integer}

\item{LocSrch_maxIter}{An integer}

\item{messageInd}{A boolean}

\item{independent.regs}{A boolean}
}
\value{
A list
}
\description{
grid.gen: generate grid for ross-validation function
}
\examples{

#####################################################################################
##### First Time Loading, Julia is Installed But Packages NEED INSTALLATION  ######
#####################################################################################
smtl_setup(path = "/Applications/Julia-1.5.app/Contents/Resources/julia/bin", installJulia = TRUE, installPackages = TRUE)
}
