% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEigen.R
\name{maxEigen}
\alias{maxEigen}
\title{maxEigen: maximum eigenvalue wrapper for Julia TSVD package. internal package use only}
\usage{
maxEigen(X, intercept = TRUE)
}
\arguments{
\item{X}{A matrix.}

\item{intercept}{A boolean.}
}
\value{
A numeric scalar of the maximum eigenvalue of provided matrix, X.
}
\description{
maxEigen: maximum eigenvalue wrapper for Julia TSVD package. internal package use only
}
