% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTaskRmse.R
\name{multiTaskRmse}
\alias{multiTaskRmse}
\title{multiTaskRmse: RMSE for multi-task problems (averaged across tasks)}
\usage{
multiTaskRmse(data, beta)
}
\arguments{
\item{data}{A matrix including outcome vector/matrix and design matrix to test RMSE on}

\item{beta}{A matrix of estimated beta coefficients where each task is in a different column}
}
\value{
Returns a scalar of average (across tasks) RMSE for predictions on data provided
}
\description{
multiTaskRmse: RMSE for multi-task problems (averaged across tasks)
}
