% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTaskRmse_MT.R
\name{multiTaskRmse_MT}
\alias{multiTaskRmse_MT}
\title{multiTaskRmse: calculate average (across tasks) RMSE for multi-label prediction problems}
\usage{
multiTaskRmse_MT(data, K = NA, beta)
}
\arguments{
\item{data}{A matrix including outcome vector/matrix and design matrix to test RMSE on}

\item{K}{An integer of number of studies/tasks}

\item{beta}{A matrix of estimated beta coefficients where each task is in a different column}
}
\value{
Returns a scalar of average (across tasks) RMSE for predictions on data provided
}
\description{
multiTaskRmse: calculate average (across tasks) RMSE for multi-label prediction problems
}
