% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reName_cv.R
\name{reName_cv}
\alias{reName_cv}
\title{reName_cv: rename output from CV. For internal package use only.}
\usage{
reName_cv(x)
}
\arguments{
\item{x}{A list (S3 class) supplied from internal sMTL functions}
}
\value{
A list (S3 class) with elements renamed. \item{best}{ A list (S3 class) with hyperparameters that achieve lowest average RMSE.} 
\item{best.1se}{ A list (S3 class) with hyperparameters associated with lowest sparsity level within 1 standard deviation of hyperparameters that achieve lowest average RMSE.}
\item{lambda_1}{ Numeric hyperparameter for L2 (ridge penalty).} \item{lambda_2}{ Numeric hyperparameter for betabar penalty.}
\item{rho}{ Integer specifying sparsity level (s).}
}
\description{
reName_cv: rename output from CV. For internal package use only.
}
