% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seReturn.R
\name{seReturn}
\alias{seReturn}
\title{seReturn: find smallest rho within 1 se of smallest cv error. For internal package use.}
\usage{
seReturn(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
Returns a dataframe that includes summary statistics to choose the best sparsity level (s) according to the 1-standard deviation rule.
}
\description{
seReturn: find smallest rho within 1 se of smallest cv error. For internal package use.
}
