% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtl.R
\name{smtl}
\alias{smtl}
\title{smtl: make model-fitting function}
\usage{
smtl(
  y,
  X,
  study = NA,
  s,
  commonSupp = FALSE,
  warmStart = TRUE,
  lambda_1 = 0,
  lambda_2 = 0,
  lambda_z = 0,
  scale = TRUE,
  maxIter = 10000,
  LocSrch_maxIter = 50,
  messageInd = TRUE,
  model = TRUE,
  independent.regs = FALSE
)
}
\arguments{
\item{y}{A numeric vector}

\item{X}{A matrix}

\item{s}{An integer}

\item{commonSupp}{A boolean}

\item{lambda_1}{A numeric vector}

\item{lambda_2}{A numeric vector}

\item{lambda_z}{A numeric vector}

\item{scale}{A boolean}

\item{maxIter}{An integer}

\item{LocSrch_maxIter}{An integer}

\item{messageInd}{A boolean}

\item{model}{A boolean}

\item{independent.regs}{A boolean}
}
\value{
A model object (list)
}
\description{
smtl: make model-fitting function
}
\examples{

#####################################################################################
##### First Time Loading, Julia is Installed and Julia Path is Known ######
#####################################################################################
smtl_setup(path = "/Applications/Julia-1.5.app/Contents/Resources/julia/bin", installJulia = FALSE, installPackages = FALSE)"

#######################################################################################################
##### If you have run smtl_setup() before, then path specification shouldn't be necessary ######
#######################################################################################################
smtl_setup(path = NULL, installJulia = FALSE, installPackages = FALSE)"

#####################################################################################
##### First Time Loading, Julia is Not Installed   ######
#####################################################################################
smtl_setup(path = NULL, installJulia = TRUE, installPackages = FALSE)"

#####################################################################################
##### First Time Loading, Julia is Installed But Packages NEED INSTALLATION  ######
#####################################################################################
smtl_setup(path = "/Applications/Julia-1.5.app/Contents/Resources/julia/bin", installJulia = TRUE, installPackages = TRUE)"
}
