% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseFn_MT.R
\name{sparseCV_MT}
\alias{sparseCV_MT}
\title{sparseCV_MT: cross-validation functions}
\usage{
sparseCV_MT(
  data,
  tune.grid,
  hoso = "hoso",
  method = "L0",
  nfolds = "K",
  juliaFnPath = NA,
  messageInd = FALSE,
  LSitr = 50,
  LSspc = 1,
  maxIter = 2500
)
}
\arguments{
\item{y}{A numeric vector}

\item{X}{A matrix}

\item{s}{An integer}
}
\description{
sparseCV_MT: cross-validation functions
}
