% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseL0Tn_iht.R
\name{sparseL0Tn_iht}
\alias{sparseL0Tn_iht}
\title{sparseCV_L0: cross-validation functions}
\usage{
sparseL0Tn_iht(
  data,
  tune.grid,
  hoso = "hoso",
  nfolds = "K",
  juliaFnPath = "/Users/gabeloewinger/Desktop/Research Final/Sparse Multi-Study/",
  trainingStudy = NA,
  messageInd = FALSE,
  LSitr = 50,
  LSspc = 1,
  maxIter = 2500
)
}
\arguments{
\item{y}{A numeric vector}

\item{X}{A matrix}

\item{s}{An integer}
}
\description{
sparseCV_L0: cross-validation functions
}
