% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.smtl.R
\name{cv.smtl}
\alias{cv.smtl}
\title{cv.smtl: cross-validation function}
\usage{
cv.smtl(
  y,
  X,
  study = NA,
  grid = NA,
  nfolds = NA,
  commonSupp = FALSE,
  multiTask = TRUE,
  lambda_1 = TRUE,
  lambda_2 = FALSE,
  lambda_z = TRUE,
  maxIter = 2500,
  LocSrch_skip = 1,
  LocSrch_maxIter = 10,
  messageInd = FALSE,
  independent.regs = FALSE
)
}
\arguments{
\item{y}{A numeric vector or matrix (for multi-label problems)}

\item{X}{A matrix}

\item{study}{An integer vector}

\item{grid}{A dataframe}

\item{commonSupp}{A boolean}

\item{multiTask}{A boolean}

\item{lambda_1}{A boolean}

\item{lambda_2}{A boolean}

\item{lambda_z}{A boolean}

\item{maxIter}{An integer}

\item{LocSrch_skip}{An integer}

\item{LocSrch_maxIter}{An integer}

\item{messageInd}{A boolean}

\item{independent.regs}{A boolean}

\item{nfold}{An integer}
}
\value{
A list
}
\description{
cv.smtl: cross-validation function
}
\examples{

#####################################################################################
##### First Time Loading, Julia is Installed But Packages NEED INSTALLATION  ######
#####################################################################################
smtl_setup(path = "/Applications/Julia-1.5.app/Contents/Resources/julia/bin", installJulia = TRUE, installPackages = TRUE)
}
