% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.gen.R
\name{grid.gen}
\alias{grid.gen}
\title{grid.gen: generate grid for cross-validation function. For internal package use only.}
\usage{
grid.gen(
  y,
  p,
  study = NA,
  lambda_1 = TRUE,
  lambda_2 = FALSE,
  lambda_z = TRUE,
  commonSupp = FALSE,
  multiTask = TRUE
)
}
\arguments{
\item{y}{A numeric vector or matrix of outcomes}

\item{p}{An integer of covariate dimension}

\item{study}{An integer vector of task IDs}

\item{lambda_1}{A boolean}

\item{lambda_2}{A boolean}

\item{lambda_z}{A boolean}

\item{commonSupp}{A boolean}

\item{multiTask}{A boolean}
}
\value{
A dataframe
}
\description{
grid.gen: generate grid for cross-validation function. For internal package use only.
}
