% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.gen.R
\name{grid.gen}
\alias{grid.gen}
\title{grid.gen: generate grid for cross-validation function}
\usage{
grid.gen(
  y,
  p,
  study = NA,
  lambda_1 = TRUE,
  lambda_2 = FALSE,
  lambda_z = TRUE,
  commonSupp = FALSE,
  multiTask = TRUE
)
}
\arguments{
\item{y}{A numeric vector or matrix}

\item{p}{An integer}

\item{study}{An integer vector}

\item{lambda_1}{A boolean}

\item{lambda_2}{A boolean}

\item{lambda_z}{A boolean}

\item{commonSupp}{A boolean}

\item{multiTask}{A boolean}
}
\value{
A dataframe
}
\description{
grid.gen: generate grid for cross-validation function
}
\examples{

#####################################################################################
##### First Time Loading, Julia is Installed But Packages NEED INSTALLATION  ######
#####################################################################################
smtl_setup(path = "/Applications/Julia-1.5.app/Contents/Resources/julia/bin", installJulia = TRUE, installPackages = TRUE)
}
