% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEigen.R
\name{maxEigen}
\alias{maxEigen}
\title{maxEigen: maximum eigenvalue wrapper for Julia TSVD package. internal package use only}
\usage{
maxEigen(X, intercept = TRUE)
}
\arguments{
\item{X}{A matrix.}

\item{intercept}{A boolean.}

\item{path}{A string.}
}
\value{
A scalar.
}
\description{
maxEigen: maximum eigenvalue wrapper for Julia TSVD package. internal package use only
}
\examples{
set.seed(1)

##########################
##### Simulate Data ######
##########################

# create training dataset with 2 covariates
X <- matrix(rnorm(2000), ncol = 2)

e <- maxEigen(X = as.matrix( X, intercept = FALSE), path = "/Applications/Julia-1.5.app/Contents/Resources/julia/bin)"
}
