% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_names.R
\name{method_nm}
\alias{method_nm}
\title{methods names: give name for printing. Internal package use only.}
\usage{
method_nm(method, multiLabel = TRUE)
}
\arguments{
\item{method}{A string}

\item{multiLabel}{A boolean}
}
\description{
methods names: give name for printing. Internal package use only.
}
