% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTaskRmse.R
\name{multiTaskRmse}
\alias{multiTaskRmse}
\title{multiTaskRmse: RMSE for multi-task}
\usage{
multiTaskRmse(data, beta)
}
\arguments{
\item{data}{A matrix}

\item{beta}{A matrix}
}
\description{
multiTaskRmse: RMSE for multi-task
}
