% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTaskRmse_MT.R
\name{multiTaskRmse_MT}
\alias{multiTaskRmse_MT}
\title{multiTaskRmse: RMSE for multi-task}
\usage{
multiTaskRmse_MT(data, K = NA, beta)
}
\arguments{
\item{data}{A matrix}

\item{K}{An integer}

\item{beta}{A matrix}
}
\description{
multiTaskRmse: RMSE for multi-task
}
