% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{predict: predict on smtl model object}
\usage{
predict(model, X, lambda_1 = NA, lambda_2 = NA, lambda_z = NA, stack = FALSE)
}
\arguments{
\item{model}{A numeric vector}

\item{X}{A matrix}

\item{lambda_1}{A optional numeric scalar. Only needed if the model object is fit on a path (multiple hyperparameterr values)}

\item{lambda_2}{A optional numeric scalar. Only needed if the model object is fit on a path (multiple hyperparameterr values)}

\item{lambda_z}{A numeric scalar}

\item{stack}{An optional boolean to use only for Domain Generalization problems.}
}
\value{
A matrix of task-specific predictions for multi-task/multi-label or for Domain Generalization problems, average and multi-study stacking predictions.
}
\description{
predict: predict on smtl model object
}
\examples{

#####################################################################################
##### First Time Loading, Julia is Installed and Julia Path is Known ######
#####################################################################################
smtl_setup(path = "/Applications/Julia-1.5.app/Contents/Resources/julia/bin", installJulia = FALSE, installPackages = FALSE)
}
