% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reName_cv.R
\name{reName_cv}
\alias{reName_cv}
\title{reName_cv: rename output from CV. For internal package use only.}
\usage{
reName_cv(x)
}
\arguments{
\item{x}{A list
@return A list with elements renamed}
}
\description{
reName_cv: rename output from CV. For internal package use only.
}
