% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seReturn.R
\name{seReturn}
\alias{seReturn}
\title{seReturn: find smallest rho within 1 se of smallest cv error}
\usage{
seReturn(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
seReturn: find smallest rho within 1 se of smallest cv error
}
