% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseL0Tn_iht.R
\name{sparseL0Tn_iht}
\alias{sparseL0Tn_iht}
\title{sparseCV_L0: cross-validation functions. For internal package use only.}
\usage{
sparseL0Tn_iht(
  data,
  tune.grid,
  hoso = "hoso",
  nfolds = "K",
  juliaFnPath = "/Users/gabeloewinger/Desktop/Research Final/Sparse Multi-Study/",
  trainingStudy = NA,
  messageInd = FALSE,
  LSitr = 50,
  LSspc = 1,
  maxIter = 2500
)
}
\arguments{
\item{data}{Matrix with outcome and design matrix}

\item{tune.grid}{A data.frame of tuning values}

\item{hoso}{String specifying tuning type}

\item{nfolds}{String or integer specifying number of folds}

\item{juliaFnPath}{String specifying path to Julia binary}

\item{trainingStudy}{Integer specifying index of training study}

\item{messageInd}{Boolean for message printing}

\item{LSitr}{Integer specifying do <LSitr> local search iterations on parameter values where we do actually do LS; NA does no local search}

\item{LSspc}{Integer specifying number of hyperparameters to conduct local search: conduct local search every <LSspc>^th iteration. NA does no local search}

\item{maxIter}{Integer specifying max iterations of coordinate descent}

\item{method}{Sting specifying regression method}
}
\description{
sparseCV_L0: cross-validation functions. For internal package use only.
}
