% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneZscale.R
\name{tuneZscale}
\alias{tuneZscale}
\title{tuneZscale: scale lambda_z depending on magnitude. For internal package use only.}
\usage{
tuneZscale(tune.grid, rhoScale)
}
\arguments{
\item{tune.grid}{A dataframe}

\item{rhoScale}{A dataframe}
}
\description{
tuneZscale: scale lambda_z depending on magnitude. For internal package use only.
}
